Add bottom nav with Timeline and Apps tabs
Onboarding dropped new users on the app list with no guidance. With the
Timeline now live-updating and owning a guiding empty state, promote it
to a peer destination via a bottom navigation bar, landing users there
by default while keeping the filterable app list and SearchView a tap
away.

- Extract TimelineFragment from ActivityTimeline (same live refresh,
  swipe refresh, and context-aware empty state with "Open an app" CTA)
  and reuse it in both ActivityMain and ActivityTimeline.
- Add BottomNavigationView to main.xml with Timeline and Apps items;
  ActivityMain swaps visibility of the Apps content vs. the Timeline
  fragment container and adjusts the toolbar (custom on/off switch
  visible only on the Apps tab; screen title shown on Timeline).
- onPrepareOptionsMenu hides search/filter/sort on the Timeline tab.
- Remove the now-redundant "Tracker activity" overflow menu item and
  the ActivityMain->ActivityTimeline launcher intent. The standalone
  ActivityTimeline still exists as a thin fragment host so the
  Insights hero card's "timeline" shortcut keeps working (it now
  switches tabs in-place when launched from ActivityMain).
- Back from Timeline returns to Apps instead of exiting.
