+++ app/src/main/res/menu/bottom_nav.xml
@@ -1,0 +1,10 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/nav_timeline"
+        android:icon="@drawable/ic_history"
+        android:title="@string/tab_timeline" />
+    <item
+        android:id="@+id/nav_apps"
+        android:icon="@drawable/ic_apps"
+        android:title="@string/tab_apps" />
+</menu>
