+++ app/src/main/res/menu/main.xml
@@ -34,6 +34,7 @@
         </menu>
     </item>
     <item
+        android:id="@+id/menu_sort"
         android:icon="@drawable/ic_sort_white_24dp"
         android:title="@string/menu_sort">
         <menu>
@@ -57,9 +58,6 @@
             </group>
         </menu>
     </item>
-    <item
-        android:id="@+id/menu_timeline"
-        android:title="@string/menu_timeline"/>
     <item
         android:id="@+id/menu_log"
         android:title="@string/menu_log"/>
